Add files via upload
This is one of my first Mobile Application Development lab, and this application uses a NavigationContainer and a DrawerNavigator to allow navigation between various screens such as Feed, Article, Home, About, and Contact. Each screen is designed to represent a distinct view within the app, showcasing how to manage navigation and pass parameters between screens in a React Native application.