Add files via upload
This React Native application was created for a lab in my Mobile Application Development class. It showcases a to-do list with navigation between a Home Screen and an About Screen.

The app employs React Navigation for routing and a MainLayout component for a consistent layout across different screens and displays key development concepts such as component-based architecture, state management, navigation between screens, and layout consistency.

It's designed to provide a functional and user-friendly experience for managing daily tasks, coupled with navigation to learn more about the project.