Add files via upload
This Java project was created for a lab in my Object-Oriented II class, it showcases the implementation of fundamental algorithms and data structures through several Java classes, demonstrating sorting, searching, and object-oriented programming principles.

It includes basic sorting algorithms with BubbleSorter, object comparison and sorting using Comparable and Comparator interfaces with the Student and StudentTest classes, and efficient searching with the BinarySearch and BinarySearchTest classes.

This collection of classes is ideal for Java beginners to grasp essential algorithms, sorting and searching techniques, and the use of interfaces for object comparisons. It also emphasizes the importance of algorithmic thinking and problem-solving in software development.