Add files via upload
This is my most current project, for my capstone class in my final semester of my Software Development diploma.

This project is a website that is built to combat the deficiencies that the SAIT Climbing Club (SCC) is experiencing with their current system. The website is used to enroll users into the SCC database. Created users have the ability to view, register for, and attend upcoming events for the club. Administrative users have the ability to create events that include an event's id, title, date, time, and description and reminders that include a reminder's id, title, and a linked event that the reminder will inform users about that are registered for that event.

The project has been created within VS Code, with JavaScript and Tailwind CSS as the programming languages, Prisma as the ORM, and MongoDB as the database that stores the user's, event's, and reminder's information.

Although the application is not fully completed yet, the project displays my skills and abilities as a software developer and a work-in-progress that will soon be a fully-fledged website.